Split specific handler tests out of test_main
