+++ binderhub/tests/test_main.py
@@ -7,32 +7,7 @@ import jwt
 import pytest
 from bs4 import BeautifulSoup
 
-from binderhub import __version__ as binder_version
-
 from .utils import async_requests
-
-
-@pytest.mark.parametrize(
-    "old_url, new_url",
-    [
-        (
-            "/repo/binderhub-ci-repos/requirements",
-            "/v2/gh/binderhub-ci-repos/requirements/master",
-        ),
-        (
-            "/repo/binderhub-ci-repos/requirements/",
-            "/v2/gh/binderhub-ci-repos/requirements/master",
-        ),
-        (
-            "/repo/binderhub-ci-repos/requirements/notebooks/index.ipynb",
-            "/v2/gh/binderhub-ci-repos/requirements/master?urlpath=%2Fnotebooks%2Findex.ipynb",
-        ),
-    ],
-)
-async def test_legacy_redirect(app, old_url, new_url):
-    r = await async_requests.get(app.url + old_url, allow_redirects=False)
-    assert r.status_code == 302
-    assert r.headers["location"] == new_url
 
 
 @pytest.mark.remote
@@ -42,26 +17,6 @@ async def test_custom_template(app):
     r = await async_requests.get(app.url)
     assert r.status_code == 200
     assert "test-template" in r.text
-
-
-@pytest.mark.remote
-async def test_versions_handler(app):
-    # Check that the about page loads
-    r = await async_requests.get(app.url + "/versions")
-    assert r.status_code == 200
-
-    data = r.json()
-    # builder_info is different for KubernetesExecutor and LocalRepo2dockerBuild
-    try:
-        import repo2docker
-
-        allowed_builder_info = [{"repo2docker-version": repo2docker.__version__}]
-    except ImportError:
-        allowed_builder_info = []
-    allowed_builder_info.append({"build_image": app.build_image})
-
-    assert data["builder_info"] in allowed_builder_info
-    assert data["binderhub"].split("+")[0] == binder_version.split("+")[0]
 
 
 @pytest.mark.parametrize(
